ModelRecommendation/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
scikit-learn==1.4.0
requests==2.28.2
pandas==2.1.4
numpy==1.26.4
pytest==7.3.1
python-consul2==0.1.5
